refactor tests to run faster
