+++ .rubocop.yml
@@ -234,8 +234,9 @@ RSpec/DescribeClass:
 RSpec/ExampleLength:
   Max: 15
   Exclude:
-    - "spec/models/institution_program_spec.rb"
-    - "spec/models/institution_builder_spec.rb"
+    - "spec/models/institution_builder/institution_program_spec.rb"
+    - "spec/models/institution_builder/institution_builder_run_add_data_spec.rb"
+    - "spec/models/institution_builder/institution_builder_run_spec.rb"
 
 # Determined to be too benign and/or numerous to justify changing
 RSpec/LeakyConstantDeclaration:
