+++ spec/models/institution_builder/institution_builder_run_add_data_spec.rb
@@ -19,59 +19,6 @@ RSpec.describe InstitutionBuilder, type: :model do
       create :crosswalk, :institution_builder
     end
 
-    context 'when successful' do
-      it 'generates a new production version' do
-        create :version
-        old_version = Version.current_production
-        described_class.run(user)
-        version = Version.current_production
-        expect(version).not_to eq(old_version)
-        expect(version.production).to be_truthy
-        expect(version).not_to be_generating
-      end
-
-      it "writes '#{CommonInstitutionBuilder::VersionGeneration::PUBLISH_COMPLETE_TEXT}' to the log" do
-        allow(Rails.logger).to receive(:info)
-        described_class.run(user)
-        expect(Rails.logger).to have_received(:info).with(/Version\sgenerated\sand\spublished/).at_least(:once)
-      end
-
-      it 'does not write "error" to the log' do
-        allow(Rails.logger).to receive(:error)
-        described_class.run(user)
-        expect(Rails.logger).to have_received(:error).with(/error/).exactly(0).times
-      end
-    end
-
-    context 'when not successful' do
-      it 'logs errors at the database level' do
-        error_message = 'There was an error occurring at the database level: BOOM!'
-        statement_invalid = ActiveRecord::StatementInvalid.new('BOOM!')
-        statement_invalid.set_backtrace(%(backtrace))
-        allow(factory_class).to receive(:add_crosswalk).and_raise(statement_invalid)
-        allow(Rails.logger).to receive(:error).with(error_message)
-        described_class.run(user)
-        expect(Rails.logger).to have_received(:error).with(error_message)
-      end
-
-      it 'logs errors at the Rails level' do
-        error_message = 'There was an error of unexpected origin: BOOM!'
-        allow(factory_class).to receive(:add_crosswalk).and_raise(StandardError, 'BOOM!')
-        allow(Rails.logger).to receive(:error).with(error_message)
-        described_class.run(user)
-        expect(Rails.logger).to have_received(:error).with(error_message)
-      end
-
-      it 'does not change the institutions or versions if not successful' do
-        allow(factory_class).to receive(:add_crosswalk).and_raise(StandardError, 'BOOM!')
-        create :version
-        version = Version.current_production
-        described_class.run(user)
-        expect(Institution.count).to be_zero
-        expect(Version.current_production).to eq(version)
-      end
-    end
-
     describe 'when initializing with Weam data' do
       let(:weam) { Weam.first }
       let(:institution) { institutions.first }
@@ -186,170 +133,6 @@ RSpec.describe InstitutionBuilder, type: :model do
 
       it 'sets eight_keys to TRUE for every eight_key record' do
         expect(institution.eight_keys).to be_truthy
-      end
-    end
-
-    describe 'when adding Accreditation data' do
-      let(:institution) { institutions.find_by(ope: accreditation_institute.ope) }
-      let!(:accreditation_institute) { create :accreditation_institute_campus }
-
-      describe 'with regards to the time frame' do
-        it 'only adds current accreditations' do
-          create :accreditation_record
-          described_class.run(user)
-
-          expect(institution.accreditation_type).not_to be_nil
-        end
-
-        it 'does not add non-current accreditations (end date is not null)' do
-          create :accreditation_record_expired
-          described_class.run(user)
-
-          expect(institution.accreditation_type).to be_nil
-        end
-      end
-
-      describe 'with regards to the accrediting authority' do
-        it 'adds institutional accreditations' do
-          create :accreditation_record
-          described_class.run(user)
-
-          expect(institution.accreditation_type).not_to be_nil
-        end
-
-        it 'does not add non-institutional accreditations (program id is not 1)' do
-          create :accreditation_record, program_id: 2
-          described_class.run(user)
-
-          expect(institution.accreditation_type).to be_nil
-        end
-      end
-
-      describe 'with regards to missing intermediate table matches' do
-        it 'does not add accreditations for institutions that do not have an ope match' do
-          weam = Weam.first
-          # rubocop:disable Rails/SkipsModelValidations
-          weam.update_columns(ope: 'junk')
-          # rubocop:enable Rails/SkipsModelValidations
-          create :accreditation_record
-          described_class.run(user)
-
-          institution = Institution.first
-          expect(institution.accreditation_type).to be_nil
-        end
-
-        it 'does not add accreditations when there is no dapip_id match' do
-          create :accreditation_record, dapip_id: -1
-          described_class.run(user)
-
-          institution = Institution.first
-          expect(institution.accreditation_type).to be_nil
-        end
-      end
-
-      describe 'the accreditation_type' do
-        AccreditationRecord::ACCREDITATIONS.each_pair do |type, agency_regex_array|
-          agency_regex_array.map(&:source).each do |agency_name|
-            it "is set to #{type} when the agency name contains '#{agency_name}'" do
-              create :accreditation_record, agency_name: 'Agency ' + agency_name
-              described_class.run(user)
-              expect(institution.accreditation_type).to eq(type)
-            end
-          end
-        end
-
-        it 'prefers national over hybrid' do
-          create :accreditation_record, agency_name: 'Biblical School'
-          create :accreditation_record, agency_name: 'Design School'
-          described_class.run(user)
-          expect(institution.accreditation_type).to eq('national')
-        end
-
-        it 'prefers regional over hybrid and national' do
-          create :accreditation_record, agency_name: 'Biblical School'
-          create :accreditation_record, agency_name: 'Middle School'
-          create :accreditation_record, agency_name: 'Design School'
-          described_class.run(user)
-          expect(institution.accreditation_type).to eq('regional')
-        end
-      end
-
-      describe 'the accreditation status' do
-        it 'is set only for the `AccreditationAction::PROBATIONARY_STATUSES`' do
-          create :accreditation_action
-          described_class.run(user)
-          expect(institution.accreditation_status).to be_nil
-        end
-
-        AccreditationAction::PROBATIONARY_STATUSES.each do |status|
-          it "is set for #{status}" do
-            create :accreditation_action, action_description: status[1..-2]
-            described_class.run(user)
-            expect(institution.accreditation_status).to eq(status[1..-2])
-          end
-        end
-
-        AccreditationAction::RESTORATIVE_STATUSES.each do |status|
-          it "with a current 'restore' action, it doesn't set the accreditation_status" do
-            create :accreditation_action, action_description: AccreditationAction::PROBATIONARY_STATUSES.first[1..-2],
-                                          action_date: '2019-01-06'
-            create :accreditation_action, action_description: status[1..-2], action_date: '2019-01-09'
-            described_class.run(user)
-            expect(institution.accreditation_status).to be_nil
-          end
-        end
-
-        it 'does not matter if an `accreditation_type` is set' do
-          create :accreditation_action_probationary
-          described_class.run(user)
-          expect(institution.accreditation_status).to be_truthy
-          expect(institution.accreditation_type).to be_nil
-        end
-
-        it 'does not matter if accreditation is current' do
-          create :accreditation_record, accreditation_end_date: '2011-01-01'
-          create :accreditation_action_probationary
-          described_class.run(user)
-          expect(institution.accreditation_status).to be_truthy
-          expect(institution.accreditation_type).to be_nil
-        end
-      end
-
-      describe 'the accreditation_action caution_flags' do
-        it 'has flags for any non-nil status' do
-          create :accreditation_action_probationary
-          described_class.run(user)
-
-          expect(CautionFlag
-                     .where({ institution_id: institution.id,
-                              source: AccreditationCautionFlag::NAME,
-                              version_id: Version.current_production.id })
-                     .count).to be > 0
-        end
-
-        it 'has no flags for any nil status' do
-          create :accreditation_action
-          described_class.run(user)
-
-          expect(CautionFlag
-                     .where({ institution_id: institution.id,
-                              source: AccreditationCautionFlag::NAME,
-                              version_id: Version.current_production.id })
-                     .count).to eq(0)
-        end
-
-        it 'concatenates `action_description` and `justification_description`' do
-          aap = create :accreditation_action_probationary
-
-          described_class.run(user)
-
-          caution_flags = CautionFlag.where({ institution_id: institution.id,
-                                              source: AccreditationCautionFlag::NAME,
-                                              version_id: Version.current_production.id }).count
-          expect(caution_flags).to be > 0
-          expect(institutions.find(institution.id).caution_flag_reason)
-            .to include(aap.action_description, aap.justification_description)
-        end
       end
     end
 
@@ -1030,175 +813,6 @@ RSpec.describe InstitutionBuilder, type: :model do
         expect(institution.longitude.to_s).to eq(census_lat_long.interpolated_longitude_latitude.split(',')[0])
         expect(institution.latitude.to_s).to eq(census_lat_long.interpolated_longitude_latitude.split(',')[1])
       end
-    end
-  end
-
-  describe '#run - with geocoding and updating geocoding from production' do
-    let(:production_version) { Version.current_production }
-    let(:institution) { create(:institution, :physical_address) }
-
-    before do
-      weam = create(:weam, :physical_address, :approved_institution)
-      weam.facility_code = '12345'
-      weam.save(validate: false)
-      institution.version = production_version
-      institution.facility_code = '12345'
-    end
-
-    it 'copies long and lat from production as part of generating' do
-      institution.longitude = 39.14
-      institution.latitude = -75.09
-      institution.save
-      described_class.run(user)
-      institution2 = Institution.last
-
-      expect(institution2.version_id).to eq(Version.current_production.id)
-      expect(institution2.longitude).to eq(39.14)
-      expect(institution2.latitude).to eq(-75.09)
-    end
-
-    it 'does not copy long and lat from production as part of generating preview if address changes' do
-      institution.longitude = 39.14
-      institution.latitude = -75.09
-      institution.physical_zip = '22222'
-      institution.save
-      described_class.run(user)
-      institution2 = Institution.last
-      expect(institution2.version_id).to eq(Version.current_production.id)
-      expect(institution2.longitude).not_to eq(39.14)
-      expect(institution2.latitude).not_to eq(-75.09)
-    end
-  end
-
-  describe 'when calculating category ratings' do
-    let(:production_version) { Version.current_production }
-    let(:institution) { create(:institution, :physical_address) }
-
-    before do
-      weam = create(:weam, :physical_address, :approved_institution)
-      weam.facility_code = '11913105'
-      weam.save(validate: false)
-      create :institution_school_rating
-      create :institution_school_rating, :second_rating
-      institution.version = production_version
-      institution.facility_code = '11913105'
-      institution.save
-    end
-
-    it 'counts number of total institution ratings' do
-      expect(InstitutionSchoolRating.count).to eq(2)
-      described_class.run(user)
-      institution2 = Institution.last
-      expect(institution2.institution_rating.institution_rating_count).to eq(2)
-    end
-
-    it 'counts the number of responses for each question' do
-      described_class.run(user)
-      institution_rating = Institution.last.institution_rating
-      1.upto(14) do |j|
-        next if j.eql?(6) # question 6 is yes/no, doesn't get used for ratings
-
-        expect(institution_rating.send("q#{j}_count")).to eq(2)
-      end
-    end
-
-    it 'averages overall institution rating' do
-      described_class.run(user)
-      institution2 = Institution.last
-      expect(institution2.institution_rating.overall_avg).to eq(1.9)
-    end
-
-    it 'calculates the average response for each question' do
-      create :institution_school_rating, :third_rating
-      described_class.run(user)
-      institution_rating = Institution.last.institution_rating
-      1.upto(14) do |j|
-        next if j.eql?(6) # question 6 is yes/no, doesn't get used for ratings
-
-        expect(institution_rating.send("q#{j}_avg")).to eq(2.0)
-      end
-    end
-
-    it 'institution rating is null without any ratings' do
-      weam = create(:weam, :public, :approved_institution)
-      weam.facility_code = '11000001'
-      weam.save(validate: false)
-
-      described_class.run(user)
-      ins = Institution.where(facility_code: '11000001').first
-      expect(ins.institution_rating).to be_nil
-    end
-
-    it 'calculates correct average ratings for every main category' do
-      described_class.run(user)
-      institution2 = Institution.last
-      expect(institution2.institution_rating.m1_avg).to eq(2.0)
-      expect(institution2.institution_rating.m2_avg).to eq(1.8)
-      expect(institution2.institution_rating.m3_avg).to eq(1.8)
-      expect(institution2.institution_rating.m4_avg).to eq(2.0)
-    end
-
-    it 'does not include null or <=0 in question counts' do
-      create :institution_school_rating, :nil_rating
-      described_class.run(user)
-      institution_rating = Institution.last.institution_rating
-      1.upto(14) do |j|
-        next if j.eql?(6) # question 6 is yes/no, doesn't get used for ratings
-
-        expect(institution_rating.send("q#{j}_count")).to eq(2) unless j.eql?(14)
-        expect(institution_rating.send("q#{j}_count")).to eq(3) if j.eql?(14)
-      end
-    end
-
-    it 'does not include null or <=0 in average calculations' do
-      create :institution_school_rating, :third_rating
-      create :institution_school_rating, :nil_rating
-
-      described_class.run(user)
-      institution_rating = Institution.last.institution_rating
-      1.upto(14) do |j|
-        next if j.eql?(6) # question 6 is yes/no, doesn't get used for ratings
-
-        expect(institution_rating.send("q#{j}_avg")).to eq(2.0) unless j.eql?(14)
-        expect(institution_rating.send("q#{j}_avg")).to eq(2.3) if j.eql?(14)
-      end
-    end
-
-    it 'treats values greater than 4 as 4 for calculating averages' do
-      create :institution_school_rating, :third_rating
-      create :institution_school_rating, :greater_than_4_rating
-
-      described_class.run(user)
-      institution_rating = Institution.last.institution_rating
-      1.upto(14) do |j|
-        next if j.eql?(6) # question 6 is yes/no, doesn't get used for ratings
-
-        expect(institution_rating.send("q#{j}_avg")).to eq(2.5)
-      end
-    end
-  end
-
-  describe 'when processing section 1015s' do
-    let(:production_version) { Version.current_production }
-    let(:institution1) { create(:institution, :section1015a) }
-    let(:institution2) { create(:institution, :section1015b) }
-
-    before do
-      create(:weam, :institution_builder, :approved_institution)
-      create(:weam, :institution_builder2, :approved_institution)
-      create(:section1015)
-      create(:section1015, :celo_n)
-
-      [institution1, institution2].each do |inst|
-        inst.version = production_version
-        inst.save
-      end
-    end
-
-    it "removes from approved institutions with celo other than 'n' " do
-      expect(Institution.approved_institutions(Version.last.id).count).to eq(2)
-      described_class.run(user)
-      expect(Institution.approved_institutions(Version.last.id).count).to eq(1)
     end
   end
 end
